+++ cloudwatch_monitoring/metadata.rb
@@ -1,15 +1,11 @@
 name             'cloudwatch_monitoring'
-maintainer       'Alexis Midon'
-maintainer_email 'alexismidon@gmail.com'
+maintainer       'Tom Skibinski'
+maintainer_email 'tomasz.skibinski@linxsys.pl'
 license          'Apache 2.0'
 description      'cloudwatch_monitoring installs the Amazon CloudWatch Monitoring Scripts for Linux - custom metrics that reports memory, swap, and disk space utilization metrics.'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version "1.2.0"
+version "1.3.0"
 
 %w{ ubuntu debian redhat centos}.each do |os|
   supports os
-end
-
-%w{cron}.each do |cookbook|
-  depends cookbook
 end