+++ cloudwatch_monitoring/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file is used to list changes made in each version of cloudwatch_monitoring.
 
+## 1.3.0:
+* remove dependency on additional cron cookbook
+* additional conditions for user, group and directory tasks
+
 ## 1.2.0:
 * query instance metadata to get instance role instead of EC2 Ohai attribute.